Show Shortcuts as part of other item categories
