+++ composeApp/src/commonMain/kotlin/io/music_assistant/client/ui/compose/home/HomeScreen.kt
@@ -5,9 +5,7 @@ package io.music_assistant.client.ui.compose.home
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.gestures.detectTapGestures
-import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxWidth
@@ -16,7 +14,6 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.LazyListState
-import androidx.compose.foundation.lazy.LazyRow
 import androidx.compose.foundation.lazy.items
 import androidx.compose.foundation.lazy.rememberLazyListState
 import androidx.compose.material.icons.Icons
@@ -30,7 +27,6 @@ import androidx.compose.material3.IconButton
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Switch
 import androidx.compose.material3.Text
-import androidx.compose.material3.TextButton
 import androidx.compose.material3.TopAppBar
 import androidx.compose.material3.TopAppBarState
 import androidx.compose.material3.rememberTopAppBarState
@@ -60,24 +56,12 @@ import io.music_assistant.client.data.model.client.items.Playlist
 import io.music_assistant.client.data.model.client.items.Podcast
 import io.music_assistant.client.data.model.client.items.PodcastEpisode
 import io.music_assistant.client.data.model.client.items.RadioStation
-import io.music_assistant.client.data.model.client.items.RecommendationFolder
 import io.music_assistant.client.data.model.client.items.Track
 import io.music_assistant.client.settings.SettingsRepository
 import io.music_assistant.client.ui.compose.common.DataState
-import io.music_assistant.client.ui.compose.common.items.AlbumWithMenu
-import io.music_assistant.client.ui.compose.common.items.ArtistWithMenu
-import io.music_assistant.client.ui.compose.common.items.AudiobookWithMenu
-import io.music_assistant.client.ui.compose.common.items.GenreWithMenu
-import io.music_assistant.client.ui.compose.common.items.LibraryActions
-import io.music_assistant.client.ui.compose.common.items.PlayHandler
-import io.music_assistant.client.ui.compose.common.items.PlaylistActions
-import io.music_assistant.client.ui.compose.common.items.PlaylistWithMenu
-import io.music_assistant.client.ui.compose.common.items.PodcastEpisodeWithMenu
-import io.music_assistant.client.ui.compose.common.items.PodcastWithMenu
-import io.music_assistant.client.ui.compose.common.items.ProgressActions
+import io.music_assistant.client.ui.compose.common.items.CategoryRow
+import io.music_assistant.client.ui.compose.common.items.ItemCategory
 import io.music_assistant.client.ui.compose.common.items.ProvideClickActions
-import io.music_assistant.client.ui.compose.common.items.RadioWithMenu
-import io.music_assistant.client.ui.compose.common.items.TrackWithMenu
 import io.music_assistant.client.ui.compose.common.items.lazyListKey
 import io.music_assistant.client.ui.compose.common.moveToEnabledBoundary
 import io.music_assistant.client.ui.compose.common.viewmodel.ActionsViewModel
@@ -87,14 +71,6 @@ import io.music_assistant.client.ui.compose.nav.TopBarLayout
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.launch
 import musicassistantclient.composeapp.generated.resources.Res
-import musicassistantclient.composeapp.generated.resources.all_albums
-import musicassistantclient.composeapp.generated.resources.all_artists
-import musicassistantclient.composeapp.generated.resources.all_audiobooks
-import musicassistantclient.composeapp.generated.resources.all_genres
-import musicassistantclient.composeapp.generated.resources.all_playlists
-import musicassistantclient.composeapp.generated.resources.all_podcasts
-import musicassistantclient.composeapp.generated.resources.all_radio
-import musicassistantclient.composeapp.generated.resources.all_tracks
 import musicassistantclient.composeapp.generated.resources.home_edit_rows
 import musicassistantclient.composeapp.generated.resources.home_save_rows
 import musicassistantclient.composeapp.generated.resources.nav_home
@@ -121,9 +97,9 @@ fun HomeScreen(
     val shortcutsState = homeScreenState.shortcuts
     var editMode by remember { mutableStateOf(false) }
 
-    val baseList = remember(recommendationsState) {
+    val baseList = remember(recommendationsState, shortcutsState) {
         if (recommendationsState is DataState.Data) {
-            recommendationsState.data
+            val recommendations = recommendationsState.data
                 .filter {
                     it.items?.any { item ->
                         item is Track ||
@@ -138,10 +114,34 @@ fun HomeScreen(
                     } == true
                 }
                 .distinctBy { it.lazyListKey() }
+                .map {
+                    ItemCategory(
+                        id = it.itemId,
+                        title = it.displayName,
+                        items = it.items.orEmpty(),
+                        lazyListKey = it.lazyListKey(),
+                        itemType = it.rowItemType,
+                    )
+                }
+
+            if (shortcutsState is DataState.Data) {
+                val shortcuts = shortcutsState.data
+                val shortcutsCategory = ItemCategory(
+                    id = "shortcuts",
+                    title = "Shortcuts",
+                    items = shortcuts.map { it.item },
+                    lazyListKey = "shortcuts",
+                    tag = HomeScreenSemantics.SHORTCUTS_ROW_TAG,
+                )
+                recommendations + shortcutsCategory
+            } else {
+                recommendations
+            }
         } else {
             emptyList()
         }
     }
+
     // Reconciled, enabled-first ordering. Authoritative for normal-mode display.
     val working = remember(baseList, homeRowsConfig) { reconcileHomeRows(baseList, homeRowsConfig) }
 
@@ -149,7 +149,7 @@ fun HomeScreen(
     // snapshotted fresh on entering edit mode.
     var items by remember { mutableStateOf(working) }
     val enabledCount = items.count { it.second }
-    val enabledByKey = remember(items) { items.associate { it.first.lazyListKey() to it.second } }
+    val enabledByKey = remember(items) { items.associate { it.first.lazyListKey to it.second } }
 
     val displayedData =
         if (editMode) items.map { it.first } else working.filter { it.second }.map { it.first }
@@ -172,7 +172,7 @@ fun HomeScreen(
                         homeScreenViewModel.saveHomeRows(
                             items.map {
                                 SettingsRepository.HomeRowPref(
-                                    it.first.itemId,
+                                    it.first.id,
                                     it.second,
                                 )
                             },
@@ -187,24 +187,24 @@ fun HomeScreen(
         },
         topAppBarState = state.topAppBarState,
     ) {
-        val rowContent: @Composable (RecommendationFolder) -> Unit = { row ->
+        val rowContent: @Composable (ItemCategory) -> Unit = {
             CategoryRow(
-                title = row.displayName,
-                rowItemType = row.rowItemType,
+                itemCategory = it,
                 onNavigateClick = onNavigateClick,
                 onPlayClick = { item, option, radio, _ ->
                     homeScreenViewModel.onPlayClick(item, option, radio)
                 },
-                onAllClick = { row.rowItemType?.let { onLibraryItemClick(it) } },
-                mediaItems = row.items.orEmpty(),
+                onAllClick = { it.itemType?.let { onLibraryItemClick(it) } },
                 playlistActions = actionsViewModel,
                 libraryActions = actionsViewModel,
                 progressActions = actionsViewModel,
                 providerIconFetcher = providerIconFetcher,
             )
         }
+
         ProvideClickActions(ClickContext.HOME) {
             LazyColumn(
+                modifier = Modifier.testTag(HomeScreenSemantics.LIST_TAG),
                 state = state.lazyListState,
                 contentPadding = contentPadding,
             ) {
@@ -218,33 +218,15 @@ fun HomeScreen(
                         }
                     }
                 } else {
-                    if (shortcutsState is DataState.Data) {
-                        item {
-                            CategoryRow(
-                                title = "Shortcuts",
-                                onNavigateClick = onNavigateClick,
-                                onPlayClick = { item, option, radio, _ ->
-                                    homeScreenViewModel.onPlayClick(item, option, radio)
-                                },
-                                mediaItems = shortcutsState.data.map { it.item },
-                                playlistActions = actionsViewModel,
-                                libraryActions = actionsViewModel,
-                                progressActions = actionsViewModel,
-                                providerIconFetcher = providerIconFetcher,
-                                rowTag = HomeScreenSemantics.SHORTCUTS_ROW_TAG,
-                            )
-                        }
-                    }
-
                     items(
                         items = displayedData,
-                        key = { it.lazyListKey() },
-                    ) { row ->
+                        key = { it.lazyListKey },
+                    ) { itemCategory ->
                         if (editMode) {
-                            val enabled = enabledByKey[row.lazyListKey()] ?: true
-                            ReorderableItem(reorderableState, key = row.lazyListKey()) {
+                            val enabled = enabledByKey[itemCategory.lazyListKey] ?: true
+                            ReorderableItem(reorderableState, key = itemCategory.lazyListKey) {
                                 Box(modifier = Modifier.fillMaxWidth()) {
-                                    rowContent(row)
+                                    rowContent(itemCategory)
                                     Box(
                                         modifier = Modifier
                                             .matchParentSize()
@@ -272,7 +254,11 @@ fun HomeScreen(
                                             enabled = !enabled || enabledCount > 1,
                                             onCheckedChange = { newEnabled ->
                                                 items =
-                                                    moveToEnabledBoundary(items, row, newEnabled)
+                                                    moveToEnabledBoundary(
+                                                        items,
+                                                        itemCategory,
+                                                        newEnabled,
+                                                    )
                                             },
                                         )
                                         Icon(
@@ -291,7 +277,7 @@ fun HomeScreen(
                                 }
                             }
                         } else {
-                            rowContent(row)
+                            rowContent(itemCategory)
                         }
                     }
                 }
@@ -330,190 +316,6 @@ private fun LandingPageTopBar(
     )
 }
 
-// --- Common UI Components ---
-
-@Composable
-fun CategoryRow(
-    title: String,
-    rowItemType: MediaType? = null,
-    onNavigateClick: (AppMediaItem) -> Unit,
-    onPlayClick: PlayHandler<AppMediaItem>,
-    onAllClick: () -> Unit = { },
-    mediaItems: List<AppMediaItem>,
-    playlistActions: PlaylistActions,
-    libraryActions: LibraryActions,
-    progressActions: ProgressActions? = null,
-    providerIconFetcher: (@Composable (Modifier, String) -> Unit),
-    rowTag: String? = null,
-) {
-    val rowListState = rememberLazyListState()
-
-    Column {
-        RowTitle(
-            title = title,
-            link = {
-                rowItemType?.let { type ->
-                    val allTitle = allItemsTitle(type)
-                    allTitle?.let {
-                        TextButton(
-                            onClick = onAllClick,
-                            contentPadding = PaddingValues(start = 4.dp, end = 4.dp),
-                        ) {
-                            Text(allTitle, style = MaterialTheme.typography.labelLarge)
-                        }
-                    }
-                }
-            },
-        )
-
-        val modifier = if (rowTag != null) {
-            Modifier.testTag(rowTag)
-        } else {
-            Modifier
-        }
-
-        LazyRow(
-            modifier = modifier,
-            state = rowListState,
-            contentPadding = PaddingValues(horizontal = 16.dp),
-            horizontalArrangement = Arrangement.spacedBy(4.dp),
-        ) {
-            items(
-                items = mediaItems,
-                key = { it.lazyListKey() },
-                contentType = { item ->
-                    when (item) {
-                        is Track -> "Track"
-                        is Artist -> "Artist"
-                        is Album -> "Album"
-                        is Playlist -> "Playlist"
-                        is Audiobook -> "Audiobook"
-                        is Podcast -> "Podcast"
-                        is PodcastEpisode -> "Episode"
-                        is RadioStation -> "RadioStation"
-                        is Genre -> "Genre"
-                        else -> "Unknown"
-                    }
-                },
-            ) { item ->
-                when (item) {
-                    is Artist -> ArtistWithMenu(
-                        item = item,
-                        onNavigateClick = onNavigateClick,
-                        onPlayOption = onPlayClick,
-                        libraryActions = libraryActions,
-                        providerIconFetcher = providerIconFetcher,
-                    )
-
-                    is Album -> AlbumWithMenu(
-                        item = item,
-                        onNavigateClick = onNavigateClick,
-                        onPlayOption = onPlayClick,
-                        playlistActions = playlistActions,
-                        libraryActions = libraryActions,
-                        providerIconFetcher = providerIconFetcher,
-                    )
-
-                    is Playlist -> PlaylistWithMenu(
-                        item = item,
-                        onNavigateClick = onNavigateClick,
-                        onPlayOption = onPlayClick,
-                        libraryActions = libraryActions,
-                        providerIconFetcher = providerIconFetcher,
-                    )
-
-                    is Podcast -> PodcastWithMenu(
-                        item = item,
-                        onNavigateClick = onNavigateClick,
-                        onPlayOption = onPlayClick,
-                        libraryActions = libraryActions,
-                        providerIconFetcher = providerIconFetcher,
-                    )
-
-                    is Track -> TrackWithMenu(
-                        item = item,
-                        onPlayOption = onPlayClick,
-                        playlistActions = playlistActions,
-                        libraryActions = libraryActions,
-                        providerIconFetcher = providerIconFetcher,
-                    )
-
-                    is PodcastEpisode -> PodcastEpisodeWithMenu(
-                        item = item,
-                        onPlayOption = onPlayClick,
-                        playlistActions = playlistActions,
-                        libraryActions = libraryActions,
-                        progressActions = progressActions,
-                        providerIconFetcher = providerIconFetcher,
-                    )
-
-                    is Audiobook -> AudiobookWithMenu(
-                        item = item,
-                        onNavigateClick = onNavigateClick,
-                        onPlayOption = onPlayClick,
-                        playlistActions = playlistActions,
-                        libraryActions = libraryActions,
-                        progressActions = progressActions,
-                        providerIconFetcher = providerIconFetcher,
-                    )
-
-                    is RadioStation -> RadioWithMenu(
-                        item = item,
-                        onPlayOption = onPlayClick,
-                        playlistActions = playlistActions,
-                        libraryActions = libraryActions,
-                        providerIconFetcher = providerIconFetcher,
-                    )
-
-                    is Genre -> GenreWithMenu(
-                        item = item,
-                        onNavigateClick = onNavigateClick,
-                        onPlayOption = onPlayClick,
-                        libraryActions = libraryActions,
-                        providerIconFetcher = providerIconFetcher,
-                    )
-
-                    else -> {}
-                }
-            }
-        }
-    }
-}
-
-@Composable
-private fun RowTitle(
-    title: String,
-    link: @Composable () -> Unit = {},
-) {
-    Row(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(horizontal = 16.dp, vertical = 8.dp),
-        verticalAlignment = Alignment.CenterVertically,
-        horizontalArrangement = Arrangement.SpaceBetween,
-    ) {
-        Text(
-            text = title,
-            style = MaterialTheme.typography.titleLarge,
-        )
-
-        link()
-    }
-}
-
-@Composable
-fun allItemsTitle(type: MediaType) = when (type) {
-    MediaType.TRACK -> stringResource(Res.string.all_tracks)
-    MediaType.ALBUM -> stringResource(Res.string.all_albums)
-    MediaType.ARTIST -> stringResource(Res.string.all_artists)
-    MediaType.PLAYLIST -> stringResource(Res.string.all_playlists)
-    MediaType.AUDIOBOOK -> stringResource(Res.string.all_audiobooks)
-    MediaType.PODCAST -> stringResource(Res.string.all_podcasts)
-    MediaType.RADIO -> stringResource(Res.string.all_radio)
-    MediaType.GENRE -> stringResource(Res.string.all_genres)
-    else -> null
-}
-
 class HomeScreenState(
     val topAppBarState: TopAppBarState,
     val lazyListState: LazyListState,
@@ -540,4 +342,5 @@ class HomeScreenState(
 
 object HomeScreenSemantics {
     const val SHORTCUTS_ROW_TAG = "ShortcutsRow"
+    const val LIST_TAG = "List"
 }
